+++ database/migrations/2024_11_20_134331_create_feedbacks_table.php
@@ -15,10 +15,10 @@ class CreateFeedbacksTable extends Migration
     {
         Schema::create('feedbacks', function (Blueprint $table) {
             $table->id();
-            $table->unsignedInteger('test_attempt_id');
+            $table->unsignedBigInteger('test_attempt_id');
             $table->text('content')->nullable();
             $table->timestamps();
-            
+
             $table->foreign('test_attempt_id')->references('id')->on('test_attempts')->onDelete('cascade');
         });
     }
